requirements.txt
anthropic>=0.48.0
GitPython>=3.1.44
openai>=1.65.0
PyGithub>=2.6.0
